move to app/lib and module RatingAdjustment
